% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{data}
\name{.my_cache}
\alias{.my_cache}
\title{Store an object in the internal package cache}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.my_cache
}
\arguments{
\item{name}{Character string — name of the object to store.}

\item{value}{Any R object — the data or model to store.}
}
\value{
Invisibly returns TRUE.
}
\description{
Store an object in the internal package cache
}
\keyword{datasets}
