% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.select.R
\name{prop.select}
\alias{prop.select}
\title{Analyse en Composantes Principales (ACP) avec extraction des variables clés}
\usage{
prop.select(df_normalized = NULL)
}
\arguments{
\item{df_normalized}{Un dataframe contenant les données normalisées après sélection.}
}
\value{
Un dataframe contenant les variables les plus importantes selon les composantes principales retenues.
}
\description{
Fonction pour réaliser une ACP, visualiser les résultats, extraire les variables les plus importantes,
et permettre à l'utilisateur de sélectionner manuellement des variables supplémentaires pertinentes.
}
