% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_data.R
\name{scale_data}
\alias{scale_data}
\title{Sélection, normalisation et exportation des choix utilisateurs}
\usage{
scale_data(donnees = NULL)
}
\arguments{
\item{donnees}{Un dataframe contenant les données à normaliser.}
}
\value{
Un dataframe avec les données normalisées après suppression des colonnes spécifiées.
}
\description{
Sélection, normalisation et exportation des choix utilisateurs
}
