% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_clust.R
\name{visual.clust}
\alias{visual.clust}
\title{Visualisation des clusters géochimiques}
\usage{
visual.clust(data)
}
\arguments{
\item{data}{Le dataframe original avec la colonne "Cluster" déjà ajoutée}
}
\value{
Un graphique ggplot2 avec les profils géochimiques colorés par cluster

Un graphique ggplot2 avec les profils géochimiques colorés par cluster
}
\description{
Crée des profils géochimiques en fonction de la profondeur, colorés selon les clusters assignés.
L'utilisateur peut définir une palette personnalisée en fournissant une couleur par cluster.

Crée des profils géochimiques en fonction de la profondeur, colorés selon les clusters assignés.
L'utilisateur peut définir une palette personnalisée en fournissant une couleur par cluster.
}
