% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charger.R
\name{charger}
\alias{charger}
\title{Load a CSV or Excel (.xlsx) file into a DataFrame.}
\usage{
charger(path = NULL, sheet = NULL, sep, fileEncoding)
}
\arguments{
\item{path}{Full path to the CSV or Excel file.}

\item{sheet}{(optional) Name or index of the Excel sheet. If not specified, the first sheet is used.}
}
\description{
Load a CSV or Excel (.xlsx) file into a DataFrame.
}
