% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrf_clust.R
\name{xrf_clust}
\alias{xrf_clust}
\title{Fonction pour normaliser les données, déterminer le nombre optimal de clusters,
et ajouter les résultats au dataframe choisi par l'utilisateur}
\usage{
xrf_clust(data)
}
\arguments{
\item{data}{Un dataframe contenant les données numériques à analyser}
}
\value{
Le dataframe avec la colonne "Cluster" ajoutée
}
\description{
Fonction pour normaliser les données, déterminer le nombre optimal de clusters,
et ajouter les résultats au dataframe choisi par l'utilisateur
}
